Call sites bug fixed (tracer still to be fixed!)

+++ trunk/testing/cct.h
@@ -16,12 +16,12 @@ struct cct_node_s {
     UINT32      counter;
     cct_node_t* first_child;
     cct_node_t* next_sibling;
-    UINT16      call_site;
+    UINT32      call_site;
 };
 
 int hcct_getenv() __attribute__((no_instrument_function));
 int hcct_init() __attribute__((no_instrument_function));
-void hcct_enter(UINT32 routine_id, UINT16 call_site) __attribute__((no_instrument_function));
+void hcct_enter(UINT32 routine_id, UINT32 call_site) __attribute__((no_instrument_function));
 void hcct_exit() __attribute__((no_instrument_function));
 void hcct_dump() __attribute__((no_instrument_function));
