+++ trunk/testing/lss-hcct.h
@@ -15,7 +15,7 @@ struct lss_hcct_node_s {
     UINT32           counter;    
     lss_hcct_node_t* first_child;
     lss_hcct_node_t* next_sibling;
-    UINT16           call_site;
+    UINT32           call_site;
     UINT16           additional_info;
     lss_hcct_node_t* parent;
     #if KEEP_EPS
@@ -25,7 +25,7 @@ struct lss_hcct_node_s {
 
 int hcct_getenv() __attribute__((no_instrument_function));
 int hcct_init() __attribute__((no_instrument_function));
-void hcct_enter(UINT32 routine_id, UINT16 call_site) __attribute__((no_instrument_function));
+void hcct_enter(UINT32 routine_id, UINT32 call_site) __attribute__((no_instrument_function));
 void hcct_exit() __attribute__((no_instrument_function));
 void hcct_dump() __attribute__((no_instrument_function));
